requirements_tf.txt
absl-py==1.4.0
ale-py==0.7.4
astunparse==1.6.3
AutoROM==0.4.2
AutoROM.accept-rom-license==0.5.4
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
efficientnet==1.0.0
flatbuffers==23.1.21
fonttools==4.38.0
gast==0.4.0
google-auth==2.16.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
gym==0.21.0
gym-notices==0.0.8
h5py==3.8.0
idna==3.4
image-classifiers==1.0.0
imageio==2.26.0
importlib-metadata==4.13.0
importlib-resources==5.10.2
keras==2.10.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keras-rl2==1.0.5
kiwisolver==1.4.4
lazy_loader==0.1
libclang==15.0.6.1
libtorrent==2.0.7
llvmlite==0.39.1
Markdown==3.4.1
markdown-it-py==2.1.0
MarkupSafe==2.1.2
matplotlib==3.6.3
mdurl==0.1.2
networkx==3.0
numba==0.56.4
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.7.0.68
opt-einsum==3.3.0
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
pip==22.3.1
protobuf==3.19.6
psutil==5.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pygame==2.1.2
pyglet==1.5.27
Pygments==2.14.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
PyWavelets==1.4.1
requests==2.28.2
requests-oauthlib==1.3.1
rich==13.3.1
rsa==4.9
scikit-image==0.20.0
scipy==1.9.1
segmentation-models==1.0.1
setuptools==65.6.3
six==1.16.0
stable-baselines3==1.7.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.1
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.30.0
termcolor==2.2.0
tflearn==0.5.0
tifffile==2023.2.28
torch==1.13.1
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.14
Werkzeug==2.2.2
wheel==0.37.1
wincertstore==0.2
wrapt==1.14.1
zipp==3.12.1
